removes the unnecessary s-expressioning of data literals
Seeing as we emit data literals in a language specific way,
and there is no need to put them in s-expressions to identify them,
it is easier to match the data literals themselves.
Not sure why we bothered with the (persistent-vector ...) etc forms.
Another advantage to letting data literals flow through is that
they can preserve meta data (types, mutability).